chore(tech): E2E tests - in all account tests, delete the account after tests have run
