+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/confirm-email/confirm-email-refresh-page.spec.js
@@ -20,6 +20,10 @@ context('Insurance - Account - Create - Confirm email page - refreshing the page
     cy.reload();
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   it('should NOT render `sent a link to` with the submitted email, because it is no longer in the session', () => {
     confirmEmailPage.weSentLinkTo().should('not.exist');
   });
