+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/confirm-email/confirm-email.spec.js
@@ -29,6 +29,10 @@ context('Insurance - Account - Create - Confirm email page - As an Exporter I wa
   let exporter;
   let expectedUrl;
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   describe('page URL and content', () => {
     before(() => {
       /**
