+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/resend-confirm-email/resend-confirm-email-go-back.spec.js
@@ -23,6 +23,10 @@ context('Insurance - Account - Create - Resend confirm email page - Go back to c
     Cypress.Cookies.preserveOnce('connect.sid');
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   let expectedUrl;
   let exporter;
