+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/resend-confirm-email/resend-confirm-email.spec.js
@@ -25,6 +25,10 @@ context('Insurance - Account - Create - Resend confirm email page - As an Export
     Cypress.Cookies.preserveOnce('connect.sid');
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   let exporter;
   let expectedUrl;
