+++ e2e-tests/cypress/e2e/journeys/insurance/account/create/your-details/account-create-your-details.spec.js
@@ -41,6 +41,10 @@ context('Insurance - Account - Create - Your details page - As an exporter, I wa
     Cypress.Cookies.preserveOnce('exip-session');
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   it('renders core page elements', () => {
     cy.corePageChecks({
       pageTitle: CONTENT_STRINGS.PAGE_TITLE,
