+++ e2e-tests/cypress/e2e/journeys/insurance/account/sign-in/enter-code/account-sign-in-enter-code.spec.js
@@ -34,6 +34,10 @@ context('Insurance - Account - Sign in - I want to sign in into my UKEF digital 
     Cypress.Cookies.preserveOnce('exip-session');
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   describe('when the account is verified', () => {
     before(() => {
       cy.verifyAccountEmail();
