+++ e2e-tests/cypress/e2e/journeys/insurance/account/sign-in/sign-in-with-second-requested-sign-in-code/sign-in-with-second-requested-sign-in-code.spec.js
@@ -50,6 +50,10 @@ context('Insurance - Account - Sign in - I want to enter the new security code s
     Cypress.Cookies.preserveOnce('exip-session');
   });
 
+  after(() => {
+    cy.deleteAccount();
+  });
+
   describe('when submitting a valid security code', () => {
     it(`should redirect to ${DASHBOARD}`, () => {
       /**
