+++ e2e-tests/cypress/support/commands.js
@@ -44,6 +44,8 @@ Cypress.Commands.add('completeAndSubmitSignInAccountForm', require('./insurance/
 Cypress.Commands.add('getExporterByEmail', require('./insurance/account/get-exporter-by-email'));
 Cypress.Commands.add('verifyAccountEmail', require('./insurance/account/verify-account-email'));
 
+Cypress.Commands.add('deleteAccount', require('./insurance/account/delete-account'));
+
 Cypress.Commands.add('completeAndSubmitPolicyTypeForm', require('./insurance/complete-and-submit-policy-type-form'));
 Cypress.Commands.add('completeAndSubmitSingleContractPolicyForm', require('./insurance/complete-and-submit-single-contract-policy-form'));
 Cypress.Commands.add('completeAndSubmitMultipleContractPolicyForm', require('./insurance/complete-and-submit-multiple-contract-policy-form'));
